Rename to state notifier
